test/inspector/runtime/await-promise-expected.txt

Tests that Runtime.awaitPromise works.

Running test: testResolvedPromise
{
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testRejectedPromise
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            objectId : 0
            type : object
            value : {
                a : 1
            }
        }
        exceptionId : 0
        lineNumber : 0
        stackTrace : {
            callFrames : [
            ]
        }
        text : Uncaught (in promise)
    }
    result : {
        type : object
        value : {
            a : 1
        }
    }
}

Running test: testRejectedPromiseWithStack
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            description : 239
            objectId : 0
            type : number
            value : 239
        }
        exceptionId : 0
        lineNumber : 0
        stackTrace : {
            callFrames : [
            ]
            parent : {
                callFrames : [
                    [0] : {
                        columnNumber : 4
                        functionName : rejectPromise
                        lineNumber : 17
                        scriptId : 0
                        url : test.js
                    }
                    [1] : {
                        columnNumber : 0
                        functionName : (anonymous)
                        lineNumber : 0
                        scriptId : 0
                        url : (empty)
                    }
                ]
                description : Promise.reject
            }
        }
        text : Uncaught (in promise)
    }
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testPendingPromise
{
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testResolvedWithoutArgsPromise
{
    result : {
        type : undefined
    }
}

Running test: testGarbageCollectedPromise
{
    code : -32000
    message : Promise was collected
}
